% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wavelet.R
\name{days_in_year}
\alias{days_in_year}
\title{Number of days in a year}
\usage{
days_in_year(year)
}
\arguments{
\item{year}{A numeric year (or vector of years).}
}
\value{
A numeric vector: 365 or 366 for each element of \code{year}.
}
\description{
Return 366 for leap years and 365 otherwise.
}
\keyword{internal}
