% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocesses the data for fitting a model}
\usage{
preprocess_data(
  dat,
  model,
  covariate,
  start_year,
  end_year,
  start_month,
  end_month,
  window_start,
  window_end,
  extra_cov = NULL
)
}
\arguments{
\item{dat}{The input data frame containing the data to be preprocessed.}

\item{model}{The model type to be used for preprocessing.}

\item{covariate}{The name of the covariate variable in the data frame.}

\item{start_year}{The start year for filtering the data.}

\item{end_year}{The end year for filtering the data.}

\item{start_month}{The start month for filtering the data.}

\item{end_month}{The end month for filtering the data.}

\item{window_start}{The start month of the window for computing covariate values.}

\item{window_end}{The end month of the window for computing covariate values.}
}
\value{
The preprocessed data frame.
}
\description{
This function preprocesses the data by filtering it based on the specified
start and end years, start and end months, and window start and end values.
It also computes smoothed versions of the population size and computes
covariate values based on the specified model.
}
\details{
The function performs the following steps:
\itemize{
\item Computes the number of years in the data frame.
\item Creates a smoothed version of the population size.
\item Computes covariate values based on the specified model and window.
\item Filters the data frame based on the specified start and end years and months.
\item Normalizes the covariate values based on the specified model.
\item Computes the population size and its derivative using the smoothed version.
}
}
\examples{
\dontrun{
# Preprocess the data using the "original" model
preprocessed_data <- preprocess_data(dat, "original", "covariate", 2000, 2020, 1, 12, 6, 9)
}
}
