% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhmalaria-package.R
\docType{package}
\name{rhmalaria-package}
\alias{rhmalaria}
\alias{rhmalaria-package}
\title{rhmalaria: Relative Humidity and Malaria Transmission Modeling Tools}
\description{
Simulation, model-fitting, and diagnostic utilities for a mechanistic
\emph{Plasmodium falciparum} transmission model relating malaria transmission
to relative humidity and temperature in Ahmedabad and Surat, India. See
the package vignettes for the full analysis pipeline.
}
\author{
\strong{Maintainer}: Leonardo Souto Ferreira \email{leonardo.souto.ferreira@gmail.com}

Authors:
\itemize{
  \item Leonardo Souto Ferreira \email{leonardo.souto.ferreira@gmail.com}
}

}
\keyword{internal}
