% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_safe.R
\name{safe_write_json}
\alias{safe_write_json}
\title{Atomically write a list/object manifest as JSON}
\usage{
safe_write_json(
  obj,
  dir = "manifests",
  prefix = "manifest",
  suffix = NULL,
  pretty = FALSE,
  auto_unbox = TRUE
)
}
\arguments{
\item{obj}{R object (usually a list) that will be converted to JSON.}

\item{dir}{Directory where the JSON will be written.}

\item{prefix}{File name prefix. Final file is \code{paste0(prefix, "_", suffix, ".json")}.}

\item{suffix}{Optional suffix; if NULL the SLURM-derived suffix is used.}

\item{pretty}{Logical, passed to \code{jsonlite::write_json()}.}

\item{auto_unbox}{Logical, passed to \code{jsonlite::write_json()}.}
}
\value{
Invisibly returns the path to the final JSON file written.
}
\description{
Serialize an R object to JSON and write it atomically to the target
directory. Uses \code{jsonlite::write_json()}; jsonlite must be available.
}
