% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_model}
\alias{simulate_model}
\title{Simulate Model}
\usage{
simulate_model(
  parameters,
  city,
  n_sims = 1,
  model = "original",
  covariate = "RH2",
  start_year = 1997,
  start_month = 1,
  end_year = 2014,
  end_month = 12,
  window_start,
  window_end,
  dataset,
  output_format = c("data.frame", "pomps"),
  extra_cov = NULL,
  model_dir = "dynamical_model/fitting",
  model_file = NULL,
  validate_only = FALSE
)
}
\arguments{
\item{parameters}{A numeric vector or data frame containing the model parameters.}

\item{city}{The name of the city.}

\item{n_sims}{The number of simulations to perform (default is 1).}

\item{model}{The model to use (default is "original").}

\item{covariate}{The covariate to use (default is "RH2").}

\item{start_year}{The starting year of the simulation (default is 1997).}

\item{start_month}{The starting month of the simulation (default is 1).}

\item{end_year}{The ending year of the simulation (default is 2014).}

\item{end_month}{The ending month of the simulation (default is 12).}

\item{window_start}{The starting point of the window for preprocessing the data.}

\item{window_end}{The ending point of the window for preprocessing the data.}

\item{dataset}{The path to the dataset file.}

\item{output_format}{The format of the output (default is "data.frame").}
}
\value{
The simulated model output in the specified format.
}
\description{
This function simulates a model using the provided parameters and dataset.
}
\details{
This function loads the dataset, preprocesses the data, and prepares the POMP model
using the provided parameters. It then performs the simulation and returns the result.
}
