% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wavelet.R
\name{yhour}
\alias{yhour}
\title{Compute the hour of the year for a date-time}
\usage{
yhour(date_hour)
}
\arguments{
\item{date_hour}{A date-time value (or vector of date-times).}
}
\value{
A numeric vector of hour-of-year values.
}
\description{
Convert a date-time into an integer hour-of-year count, 0-indexed from
midnight on January 1st of that year (i.e. Jan 1st 00:00 is hour 0).
}
\keyword{internal}
